requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
celery>=5.3.0
redis>=5.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.11.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
jinja2>=3.1.0
openpyxl>=3.1.0
pandas>=2.0.0
httpx>=0.24.0
python-multipart>=0.0.6
loguru>=0.7.0
tenacity>=8.2.0
python-dotenv>=1.0.0
zhipuai>=2.0.0
plotly>=5.15.0
scikit-learn>=1.3.0
numpy>=1.24.0
pyjwt>=2.8.0
sniffio>=1.3.0
